skynetv2/requirements.txt
aiohttp>=3.9.0
openai>=1.0.0
anthropic>=0.18.0
groq>=0.4.1
google-generativeai>=0.3.0
